Use `ws` WebSocket when options are provided (#2968)
The current 3.7-dev implementation does not pass headers to the WebSocket when utilizing the globalThis.WebSocket. This causes issues when initially upgrading to this version of gremlin (or upgrading to Node 22 when the global WebSocket class was enabled by default). This change ensures that we use the ws WebSocket implementation if ws specific options are passed along.