TINKERPOP-3017 Added AnonymousTraversalSource.with() (#3073)
Replaces withEmbedded() and withRemote() and removes the long deprecated withGraph(). A single unified with() removes the need to check the object being given to the AnonymousTraversalSource to decide which method to use. Can't remember why that split was used in the first place. This should make it easier to explain how to use the AnonymousTraversalSource in the first place, requires a bit less typing and simply streamlines the API.

Co-authored-by: Stephen Mallette <stepmall@amazon.com>