TINKERPOP-3167 Fixed NPE when borrowing a connection
There is a chance that the Client.hostConnectionPools can land in a state where it has zero hosts present which leads to the returned ConnectionPool itself to be null. Borrowing a Connection then leads to a NullPointerException.

By preferring a check during init of actual connection pool creation the Client avoids a state where it could grab a null pool. This change doesn't address some of the synchronization issues that likely contributed to allowing the NPE in the first place, but there doesn't seem to be evidence that having an inconsistent state here can lead to issues outside of init().