Only include STJ on net standard (#3123)
Adjusts the inclusion of System.Text.Json so that it is only included on net standard and not for the newer frameworks ie net 6.

The benefit is that it reduces the risk of conflicts which has occurred with other libraries, enables quicker remediation to CVE'S in STJ when end application deployed being dependent on framework as windows update can update the framework.