TINKERPOP-3141
Account for the scenario of dropping and re-adding a vertex within the same transaction by adding a new unmarkDeleted method to the TinkerElementContainer and calling unmarkDeleted from addVertex if it is detected that the vertex had been previously deleted. Added logic to reference the previously deleted vertex version when adding the vertex back to avoid transaction conflict error.