TINKERPOP-3023 Expand type syntax in grammar - Remove Vertex - TP4 (#3138)
This is a followup to #3133 with changes exclusively for TP4. The primary difference in TP4 is that bytecode has been replaced with GremlinLang, which means that the use of Vertex in V(), from(), and to() is now sugar added by GraphTraversal. These vertices must be converted to their ids when passed into GremlinLang.

---------

Co-authored-by: Peter Tribe <peter.tribe@improving.com>