TINKERPOP-3178 choose semantics consistency improvements
Only match on first option for switch semantics. Introduced Pick.unproductive when the choice ends up being unproductive. Pass through unproductive choices for switch semantics so that they are like if-else. Introduce ChooseSemantics enum to inform on whether the step is in if-else or switch form. Lots of documentation improvements.