Transaction reads should not interfere with add/update/delete (#3163)
https://issues.apache.org/jira/browse/TINKERPOP-3142

Changed TinkerElementContainer to no longer consider element reads as being 'used in a transaction' which should be reserved for element add/update/delete only. This is appropriate for TinkerTransactionGraph because the isolation level is 'read committed'. With this isolation level, we do not need to protect against 'unrepeatable reads' or 'phantom reads'.

Prior to this change, a read-only thread could cause a 'Conflict: element modified in another transaction' error in a separate thread which was attempting add/drop/update.