Deparameterize feature tests (#3186)
Inline the majority of parameters in feature test scenarios for steps
such as inject() and withSideEffect(). This is largely in preparation
for the upcoming work on GValue parameterization in gremlin-lang.

Parameters are largely unnecessary in the feature tests since they
have been migrated from gremlin-groovy to gremlin-lang. A new bit
of gherkin syntax is being added to directly set side effects without
relying on query parameters which enables adding dynamic values such
as vertex ids into side effects.

This change also exposed a bug in the Go translator, which has been
resolved as well.

Co-authored-by: xiazcy <xiazcy@gmail.com>