GValue and Antlr Parsing Cache (#3157)
This change introduces GValue, which is essentially a wrapper around a query parameter name and value. The primary goal of GValues is to persist the parameterized state of a traversal through strategy execution. This enables post-strategy Traversals to cached, and for parameterized values to remain mutable after a Traversal is retrieved from the cache.

These changes introduce new "StepContract" interfaces and "GValue Placeholder" steps for parameterizable steps. A GValue placeholder is a non-executable placeholder step which may store its arguments as GValues. There is a new finalization strategy GValueReductionStrategy which is intended to be applied after a traversal has been stored in the cache which "reduces" all placeholder steps to their concrete non-parameterized form. The primary intention of this structure is to prevent GValues from unintentionally leaking into step semantics or traversal results.

Some optimization strategies such as EarlyLimitStrategy may alter a Traversal based on a specific argument value in such a way that is not generalizable for all substitutions of that parameter. To guard against this, each traversal is given a GValueManager which tracks if a certain variable has been "pinned". A traversal with a pinned variable is only valid for the current values of that variable, and does not allow for substitution of the pinned variable.

Also included is an optional new cache in GremlinLangScriptEngine which can be used to save on antlr parsing time which is built using the GValue improvements.

---------

Co-authored-by: Stephen Mallette <stepmall@amazon.com>
Co-authored-by: xiazcy <xiazcy@gmail.com>
Co-authored-by: andreachild <andrea.child@improving.com>