Changed detatchment logic after TINKERPOP-3182
Seems to cause problems once TINKERPOP-3182 went to master branch, but basically the way it's written now leaves things with the assumption of a homogeneous result (or at least an aggregate). That is a big assumption that seems to already be breaking on master with some of the tests. CTR