[TINKERPOP-3173] Simplify comparability semantics (#3195)
Replaces the old system of ternary boolean semantics with simple binary
semantics. The triggers for "ERROR" states from illegal comparisons are
unchanged (typically comparisons with NaN or between incomparable types
such as String and int). The difference now is that instead of the ERROR
being propagated according to ternary logic semantics until a reduction
point is reached, the error now immediately returns a value of FALSE.

This will be most visible in expressions which include negations. Prior
to this change, `g.inject(NaN).not(is(1))` would produce no results as
!(NaN == 1) -> !(ERROR) -> ERROR -> traverser is filtered out. After
this change, the same traversal will return NaN as the same expression
now evaluates as !(NaN == 1) -> !(FALSE) -> TRUE -> traverser is not
filtered.

This commit also introduces NotP to model a negated P, as the old
method of getting the complement of the PBiPredicate may handle edge
cases incorrectly:

!(1 < NaN) != (1 >= NaN) -> !(FALSE) != (FALSE)