TINKERPOP-3199 Updated build system to work with setuptools>=60
Moved from setup.py and setup.cfg to pyproject.toml to conform to
current Python build setup. Version needs to be manually set now as
scripts aren't allowed to run during build process. Decided this was
best for now but could move to setuptools-scm later to make it more
automated like it was before.