Clean up gremlin-python docker-compose after move to newer setuptools CTR
Since pip is used instead of setup.py, the dependencies are installed
directly from the pyproject.toml. radish is set to a newer version as
the older version isn't able to read all the feature tags properly. This
version was already used in practice before as it was picked up at
runtime rather than the one specified in the setup.py.