Remove GenericArgument from from() and to().
from() and to() will now only accept Strings (alias for __.select()) and Traversals.
If users wish to pass a vertex id, they may embed it in __.V() or __.constant().

The primary motivation here is to ensure consistent behaviour accross providers regardless
of if their id types are numerics, strings, or other.