Prevent inject in repeat (#3253)
Modified StandardVerificationStrategy to throw error if inject() is used inside repeat() as the injections would be exhausted after the first iteration now that the RepeatUnrollStrategy no longer auto-unrolls such traversals. It is possible to use union() as an alternative step if the old behaviour is desired although there are limited use cases for inject() in repeat(). If it is determined there is a need to have injections per iteration it is possible to implement a proper solution for this in the future which does not incorrectly rely on a strategy after-effect.