Remove GenericArgument from from() and to() (#3252)
from() and to() will now only accept Strings (alias for __.select()) and Traversals. If users wish to pass a vertex id, they may embed it in __.V() or __.constant().

The primary motivation here is to ensure consistent behaviour accross providers regardless of if their id types are numerics, strings, or other.

This change is a followup to an unintended consequence stemming from https://github.com/apache/tinkerpop/pull/3133. The issue with directly allowing ID's to be passed into from/to, it creates a weird asymmetry for graphs which use Strings as their ids. For example g.addE().from(1).to(2) would work fine and create an edge from V[1] to V[2], however if the graph used String ids, the equivalent traversal (g.addE().from("a").to("b")) does not work, as the Strings are interpreted as path labels (converted to __.select("a")) instead of vertex ids. There are some ugly workarounds which can partially mitigate this issue, but overall there is no generalized fix which allows all traversals to behave the same regardless of the vertex id type.

---------

Co-authored-by: andreachild <andrea.child@improving.com>