Fix sample in repeat computer algorithm to retain sample size (#3273)
Fixed SampleGlobalStep computer algorithm to return the correct sample size when used in repeat by retaining only the traversers which have passed through all loops when merging traverser sets. Prior to this fix the traverser sets would be merged together and traversers which were excluded from the sample would potentially be added to the sample in the next loop.