Add Support for Node 22 and 24 (#3119)
Enables support for Node 22 and 24, and expands the github action test infrastructure to test all of node 20, 22, and 24. This allows us to ensure support for all currently active LTS versions of Node, instead of pinning our support to a single major version at a time.

Note that there was an issue related to the websocket library handling in node 22. As of node 22, the builtin global websockets does not allow for customized user-agent and compression headers. To limit the impact of this issue, global websockets is only used if both the user agent and compression are disabled, otherwise we fallback to the external ws library.


---------

Co-authored-by: xiazcy <xiazcy@gmail.com>